+++ NEWS.md
@@ -1,6 +1,8 @@
 # logitr 0.2.1
 
 - Added `altIDName` argument to `predictChoices()` and `predictProbs()` to preserve the row order of predictions for each alternative in each set of alternatives. Closes issue #13.
+- Fixed bug in data encoding where random parameter names were not aligned with encoded data.
+- Added input checks for all predict functions.
 
 # logitr 0.2.0